requirements.txt
pandas>=2.0.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
openai>=1.12.0 